requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
django==3.2.12
djangorestframework==3.12.4
    # via -r requirements.in
pytz==2023.3
    # via djangorestframework
sqlparse==0.4.4
    # via django
typing-extensions==4.6.3
    # via asgiref
tzdata==2023.3
    # via django
